installer: refac setup as a list of paths
