+++ gui/src/installer/message.rs
@@ -110,8 +110,7 @@ pub enum InternalBitcoindMsg {
 #[derive(Debug, Clone)]
 pub enum DefineDescriptor {
     ImportDescriptor(String),
-    PrimaryPath(DefinePath),
-    RecoveryPath(usize, DefinePath),
+    Path(usize, DefinePath),
     AddRecoveryPath,
     KeyModal(ImportKeyModal),
     SequenceModal(SequenceModal),
